ADD Config Hot Reload - partially working
Added a Filewatcher System, that monitors if the config file was
changed. Classes can register callbacks to every key in the config now,
so that if the config changes, the app can react accordingly.

Works Fully for Fullscreen mode and Window Renaming. Resizing and
setting the window to be resizable is not working atm